fix: avoid multiple decimal values computations in calculateAverageValue
Avoid computations such the direct sum of multiple decimal values.

The binary format used by the computers to represent numbers, cannot represent all decimal values. A great part of them are rounded to the nearest number that can be represented.

If you keep operating on values that cannot be exactly represented you'll end up adding more and more margin of error for the final computation.

In this specific scenario I wanted to calculate the average value from a serie of values provided. To avoid the problem described above you can ensure that you're operating on integer values ( can be exactly represented in binary format ) by multiplying each value provided by a multiple of 10.

Before returning the result that can be a decimal number you have to divide by the multiple of 10 you previously choosed.